# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Player": {
			"properties": {
				"player_id": {
					"type": "number",
					"format": "double"
				},
				"year_played": {
					"type": "string"
				},
				"first_name": {
					"type": "string"
				},
				"last_name": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"pos": {
					"type": "string"
				},
				"num": {
					"type": "string"
				},
				"year": {
					"type": "string"
				},
				"height": {
					"type": "string"
				},
				"height_inches": {
					"type": "number",
					"format": "double"
				},
				"weight": {
					"type": "number",
					"format": "double"
				},
				"hometown": {
					"type": "string"
				},
				"last_school": {
					"type": "string"
				},
				"team": {
					"$ref": "#/definitions/Team"
				},
				"allPurposeRunning": {
					"$ref": "#/definitions/AllPurposeRunning"
				},
				"fumbleReturns": {
					"$ref": "#/definitions/FumbleReturns"
				},
				"interceptions": {
					"$ref": "#/definitions/Interceptions"
				},
				"kickoffReturn": {
					"$ref": "#/definitions/KickoffReturn"
				},
				"kickoffs": {
					"$ref": "#/definitions/Kickoffs"
				},
				"miscDefense": {
					"$ref": "#/definitions/MiscDefense"
				},
				"passing": {
					"$ref": "#/definitions/Passing"
				},
				"placeKicking": {
					"$ref": "#/definitions/PlaceKicking"
				},
				"puntReturn": {
					"$ref": "#/definitions/PuntReturn"
				},
				"punting": {
					"$ref": "#/definitions/Punting"
				},
				"receiving": {
					"$ref": "#/definitions/Receiving"
				},
				"rushing": {
					"$ref": "#/definitions/Rushing"
				},
				"sacks": {
					"$ref": "#/definitions/Sacks"
				},
				"scoring": {
					"$ref": "#/definitions/Scoring"
				},
				"tackles": {
					"$ref": "#/definitions/Tackles"
				},
				"tacklesForLoss": {
					"$ref": "#/definitions/TacklesForLoss"
				},
				"totalOffense": {
					"$ref": "#/definitions/TotalOffense"
				},
				"yardsFromScrimmage": {
					"$ref": "#/definitions/YardsFromScrimmage"
				}
			},
			"required": [
				"player_id",
				"year_played",
				"first_name",
				"last_name",
				"name",
				"pos",
				"num",
				"year",
				"height",
				"height_inches",
				"weight",
				"hometown",
				"last_school",
				"team",
				"allPurposeRunning",
				"fumbleReturns",
				"interceptions",
				"kickoffReturn",
				"kickoffs",
				"miscDefense",
				"passing",
				"placeKicking",
				"puntReturn",
				"punting",
				"receiving",
				"rushing",
				"sacks",
				"scoring",
				"tackles",
				"tacklesForLoss",
				"totalOffense",
				"yardsFromScrimmage"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Team": {
			"properties": {
				"year": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"players": {
					"items": {
						"$ref": "#/definitions/Player"
					},
					"type": "array"
				},
				"conference": {
					"$ref": "#/definitions/Conference"
				}
			},
			"required": [
				"year",
				"name",
				"players",
				"conference"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Conference": {
			"properties": {
				"name": {
					"type": "string"
				},
				"year": {
					"type": "string"
				},
				"teams": {
					"items": {
						"$ref": "#/definitions/Team"
					},
					"type": "array"
				}
			},
			"required": [
				"name",
				"year",
				"teams"
			],
			"type": "object",
			"additionalProperties": true
		},
		"AllPurposeRunning": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"IntRet": {
					"type": "number",
					"format": "double"
				},
				"KickRet": {
					"type": "number",
					"format": "double"
				},
				"Plays": {
					"type": "number",
					"format": "double"
				},
				"PuntRet": {
					"type": "number",
					"format": "double"
				},
				"Recv": {
					"type": "number",
					"format": "double"
				},
				"Rush": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TotalYards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"YardsPlay": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"IntRet",
				"KickRet",
				"Plays",
				"PuntRet",
				"Recv",
				"Rush",
				"Split",
				"TotalYards",
				"YardsG",
				"YardsPlay",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"FumbleReturns": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "number",
					"format": "double"
				},
				"FumRet": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"FumRet",
				"G",
				"Split",
				"TD",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Interceptions": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Int": {
					"type": "number",
					"format": "double"
				},
				"IntG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"Int",
				"IntG",
				"Split",
				"TD",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"KickoffReturn": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Ret": {
					"type": "number",
					"format": "double"
				},
				"RetG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Ret",
				"RetG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Kickoffs": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Kickoffs": {
					"type": "number",
					"format": "double"
				},
				"Onside": {
					"type": "number",
					"format": "double"
				},
				"OutOfBounds": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"Touchback": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Kickoffs",
				"Onside",
				"OutOfBounds",
				"Split",
				"Touchback",
				"Yards",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"MiscDefense": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"FumblesForced": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"KicksPuntsBlocked": {
					"type": "number",
					"format": "double"
				},
				"PassesBrokenUp": {
					"type": "number",
					"format": "double"
				},
				"QBHurries": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"FumblesForced",
				"G",
				"KicksPuntsBlocked",
				"PassesBrokenUp",
				"QBHurries",
				"Split",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Passing": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Att": {
					"type": "number",
					"format": "double"
				},
				"AttG": {
					"type": "number",
					"format": "double"
				},
				"Comp": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Int": {
					"type": "number",
					"format": "double"
				},
				"Pct": {
					"type": "number",
					"format": "double"
				},
				"Rating": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsAtt": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Att",
				"AttG",
				"Comp",
				"G",
				"Int",
				"Pct",
				"Rating",
				"Split",
				"TD",
				"Yards",
				"YardsAtt",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"PlaceKicking": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Att": {
					"type": "number",
					"format": "double"
				},
				"AttG": {
					"type": "number",
					"format": "double"
				},
				"ExtraPoint": {
					"type": "number",
					"format": "double"
				},
				"FieldGoal": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Made": {
					"type": "number",
					"format": "double"
				},
				"MadeG": {
					"type": "number",
					"format": "double"
				},
				"Pct": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Att",
				"AttG",
				"ExtraPoint",
				"FieldGoal",
				"G",
				"Made",
				"MadeG",
				"Pct",
				"Split",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"PuntReturn": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Ret": {
					"type": "number",
					"format": "double"
				},
				"RetG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Ret",
				"RetG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Punting": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Punts": {
					"type": "number",
					"format": "double"
				},
				"PuntsG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Punts",
				"PuntsG",
				"Split",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Receiving": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Rec": {
					"type": "number",
					"format": "double"
				},
				"RecG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Avg",
				"G",
				"Rec",
				"RecG",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Rushing": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Att": {
					"type": "number",
					"format": "double"
				},
				"AttG": {
					"type": "number",
					"format": "double"
				},
				"Avg": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"Yards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Att",
				"AttG",
				"Avg",
				"G",
				"Split",
				"TD",
				"Yards",
				"YardsG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Sacks": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"SackYards": {
					"type": "number",
					"format": "double"
				},
				"Sacks": {
					"type": "number",
					"format": "double"
				},
				"SacksG": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"SackYards",
				"Sacks",
				"SacksG",
				"Split",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Scoring": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"XP": {
					"type": "number",
					"format": "double"
				},
				"FG": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Points": {
					"type": "number",
					"format": "double"
				},
				"PointsG": {
					"type": "number",
					"format": "double"
				},
				"Safety": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"XP",
				"FG",
				"G",
				"Points",
				"PointsG",
				"Safety",
				"Split",
				"TD",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"Tackles": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"Assisted": {
					"type": "number",
					"format": "double"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Solo": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"Total": {
					"type": "number",
					"format": "double"
				},
				"TotalG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"Assisted",
				"G",
				"Solo",
				"Split",
				"Total",
				"TotalG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"TacklesForLoss": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TFL": {
					"type": "number",
					"format": "double"
				},
				"TFLYards": {
					"type": "number",
					"format": "double"
				},
				"TFLG": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"Split",
				"TFL",
				"TFLYards",
				"TFLG",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"TotalOffense": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"PassYards": {
					"type": "number",
					"format": "double"
				},
				"Plays": {
					"type": "number",
					"format": "double"
				},
				"RushYards": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TotalYards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"YardsPlay": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"PassYards",
				"Plays",
				"RushYards",
				"Split",
				"TotalYards",
				"YardsG",
				"YardsPlay",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
		"YardsFromScrimmage": {
			"properties": {
				"playerid": {
					"type": "number",
					"format": "double"
				},
				"yearplayed": {
					"type": "string"
				},
				"G": {
					"type": "number",
					"format": "double"
				},
				"Plays": {
					"type": "number",
					"format": "double"
				},
				"RecvYards": {
					"type": "number",
					"format": "double"
				},
				"RushYards": {
					"type": "number",
					"format": "double"
				},
				"Split": {
					"type": "string"
				},
				"TD": {
					"type": "number",
					"format": "double"
				},
				"TotalYards": {
					"type": "number",
					"format": "double"
				},
				"YardsG": {
					"type": "number",
					"format": "double"
				},
				"YardsPlay": {
					"type": "number",
					"format": "double"
				},
				"player": {
					"$ref": "#/definitions/Player"
				}
			},
			"required": [
				"playerid",
				"yearplayed",
				"G",
				"Plays",
				"RecvYards",
				"RushYards",
				"Split",
				"TD",
				"TotalYards",
				"YardsG",
				"YardsPlay",
				"player"
			],
			"type": "object",
			"additionalProperties": true
		},
...
		"PaginatedTeam": {
			"properties": {
				"teams": {
					"items": {
						"$ref": "#/definitions/Team"
					},
					"type": "array"
				},
				"count": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"teams",
				"count"
			],
			"type": "object",
			"additionalProperties": true
		},
...
	},
	"info": {
		"title": "new-typeorm-project",
		"version": "0.0.1",
		"description": "Awesome project developed with TypeORM."
	},
	"paths": {
...
		"/teams/{conference_year}/{conference_name}": {
			"get": {
				"operationId": "GetTeamsForConference",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Fetched",
						"schema": {
							"$ref": "#/definitions/PaginatedTeam"
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "conference_year",
						"required": true,
						"type": "string"
					},
					{
						"in": "path",
						"name": "conference_name",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"collectionFormat": "multi",
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/teams/{year}": {
			"get": {
				"operationId": "GetTeamsForYear",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginatedTeam"
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "year",
						"required": true,
						"type": "string"
					},
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"collectionFormat": "multi",
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}